refactor(routes): Rename regions to gc for golf courses pages
