+++ apps/nextjs/src/app/(routes)/gc/[z_code]/[zs_code]/page.tsx
