+++ apps/nextjs/src/app/(routes)/gc/[z_code]/page.tsx
