+++ apps/nextjs/src/app/(routes)/gc/layout.tsx
