+++ apps/nextjs/src/app/(routes)/gc/page.tsx
