+++ apps/nextjs/src/app/(routes)/golf-courses/[slug]/page.tsx
@@ -12,7 +12,7 @@ import type { GolfCourse } from "@/types";
 import { Loader2 } from "lucide-react";
 
 import BreadcrumbNav from "../../../../components/nav/breadcrumb-nav";
-import Nav from "../../regions/nav";
+import Nav from "../../gc/nav";
 import AdBanner from "./_components/ad-banner";
 import CourseCommonInfo from "./_components/course-common-info";
 import CourseDetailInfo from "./_components/course-detail-info";
